# perf: improve gpx and tcx parsing speed