--- a/src/wasm/activity-service/go.mod
+++ b/src/wasm/activity-service/go.mod
@@ -3,11 +3,12 @@ module github.com/openivity/activity-service
 go 1.20
 
 require (
-	github.com/muktihari/fit v0.19.1
-	golang.org/x/text v0.15.0
+	github.com/muktihari/fit v0.20.1
+	github.com/muktihari/xmltokenizer v0.0.0-20240614025545-7e2f70738015
+	golang.org/x/text v0.16.0
 )
 
 require (
 	github.com/google/go-cmp v0.6.0
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
+	golang.org/x/exp v0.0.0-20240613232115-7f521ea00fb8
 )