Go standard library for parsing xml is slow, now use https://github.com/muktihari/xmltokenizer instead.

Benchmark decoding 194km ride:
| Format | stdlib's xml | xmltokenizer | delta |
|---------|------------|--------------|------|
| GPX | 2823 ms   | 756 ms  |  -273%  |
| TCX | 3515 ms   | 976 ms  |  -260%  |